pyproject.toml
[tool.poetry]
name = "web-crawler"
version = "0.1.0"
description = "A simple web crawler"
authors = ["ash <ashu.banerjee.2289@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
beautifulsoup4 = "4.10.0"
requests = "^2.27.1"
PyYAML = "^6.0"
opnieuw = "^1.1.0"
argparse = "^1.4.0"
coverage = "^6.3.2"
validators = "^0.18.2"
flake8 = "^4.0.1"

[tool.poetry.dev-dependencies]
pytest = "^7.1"
pytest-mock = "^3.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
